+++ packages/design-system/src/interfaces/index.d.ts
@@ -4,3 +4,4 @@ export * from './buttons';
 export * from './helpers';
 export * from './tooltip';
 export * from './forms';
+export * from './modals';